ui: minor adjustments to social sharing
